# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Apex API",
    "description": "API de Apex — pipeline d'ingestion, recherche à facettes et espace client de Studio Chicane.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/shootings/{shooting_id}/engagements": {
      "get": {
        "tags": [
          "shootings"
        ],
        "summary": "Engagements du shooting",
        "operationId": "list_engagements_api_v1_shootings__shooting_id__engagements_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shooting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shooting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EngagementOut"
                  },
                  "title": "Response List Engagements Api V1 Shootings  Shooting Id  Engagements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "shootings"
        ],
        "summary": "Créer un engagement",
        "operationId": "create_engagement_api_v1_shootings__shooting_id__engagements_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "shooting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shooting Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/engagements/{engagement_id}": {
      "patch": {
        "tags": [
          "engagements"
        ],
        "summary": "Modifier un engagement",
        "operationId": "patch_engagement_api_v1_engagements__engagement_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Engagement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EngagementPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngagementOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "engagements"
        ],
        "summary": "Supprimer un engagement",
        "operationId": "delete_engagement_api_v1_engagements__engagement_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Engagement Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/media/{media_id}/engagements/{engagement_id}": {
      "delete": {
        "tags": [
          "media"
        ],
        "summary": "Retirer un rattachement (J2)",
        "operationId": "delete_media_engagement_api_v1_media__media_id__engagements__engagement_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "media_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Media Id"
            }
          },
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Engagement Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/queue/stats": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "État de la file",
        "operationId": "queue_stats_api_v1_queue_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/jobs/tick": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Drainer la file jusqu'à épuisement ou budget de temps (worker tiré, §3-E.7)",
        "operationId": "jobs_tick_api_v1_jobs_tick_post",
        "parameters": [
          {
            "name": "X-Worker-Secret",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Worker-Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EngagementCreate": {
        "properties": {
          "car_number": {
            "type": "string",
            "title": "Car Number"
          },
          "driver_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "car_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Car Model"
          }
        },
        "type": "object",
        "required": [
          "car_number"
        ],
        "title": "EngagementCreate"
      },
...
      "EngagementOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "shooting_id": {
            "type": "integer",
            "title": "Shooting Id"
          },
          "car_number": {
            "type": "string",
            "title": "Car Number"
          },
          "driver_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "car_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Car Model"
          }
        },
        "type": "object",
        "required": [
          "id",
          "shooting_id",
          "car_number",
          "driver_id",
          "team_id",
          "client_id",
          "car_model"
        ],
        "title": "EngagementOut"
      },
      "EngagementPatch": {
        "properties": {
          "car_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Car Number"
          },
          "driver_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "car_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Car Model"
          }
        },
        "type": "object",
        "title": "EngagementPatch"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueueStats": {
        "properties": {
          "pending": {
            "type": "integer",
            "title": "Pending"
          },
          "running": {
            "type": "integer",
            "title": "Running"
          },
          "dead": {
            "type": "integer",
            "title": "Dead"
          },
          "oldest_pending_age_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Pending Age S"
          }
        },
        "type": "object",
        "required": [
          "pending",
          "running",
          "dead",
          "oldest_pending_age_s"
        ],
        "title": "QueueStats"
      },
...
      "TickResponse": {
        "properties": {
          "claimed": {
            "type": "integer",
            "title": "Claimed"
          },
          "done": {
            "type": "integer",
            "title": "Done"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          }
        },
        "type": "object",
        "required": [
          "claimed",
          "done",
          "failed",
          "remaining"
        ],
        "title": "TickResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}